[MOVE] 카테고리 조회 request DTO 통일 위한 패키지 변경
